AddedForeignKeyToSpeakerAndTopicToEventsTable

+++ TedEx/TedEx/ViewModels/EventFormViewModel.cs
@@ -8,7 +8,7 @@ namespace TedEx.ViewModels
         public string Venue { get; set; }
         public string Date { get; set; }
         public string Time { get; set; }
-        public int Topic { get; set; }
+        public byte Topic { get; set; }
         public IEnumerable<Topic> Topics { get; set; }
     }
 }
